funções de adicionar e excluir contatos

+++ phpavancado/crud/index.php
@@ -20,16 +20,23 @@ $contato = new Contato();
     </thead>
     <tbody>
         <?php $lista = $contato->getAll(); ?>
-        <?php foreach ($lista as $contato) : ?>
+        <?php
+        if (count($lista) > 0) : ?>
+            <?php foreach ($lista as $contato) : ?>
+                <tr>
+                    <td><?= $contato['id']; ?></td>
+                    <td><?= $contato['nome']; ?></td>
+                    <td><?= $contato['email']; ?></td>
+                    <td>
+                        <a href="editar.php?id=<?= $contato['id']; ?>">[ EDITAR ]</a>
+                        <a href="excluir.php?id=<?= $contato['id']; ?>">[ EXCLUIR ]</a>
+                    </td>
+                </tr>
+            <?php endforeach; ?>
+        <?php else : ?>
             <tr>
-                <td><?= $contato['id']; ?></td>
-                <td><?= $contato['nome']; ?></td>
-                <td><?= $contato['email']; ?></td>
-                <td>
-                    <a href="editar.php?id=<?= $contato['id']; ?>">[ EDITAR ]</a>
-                    <a href="excluir.php?id=<?= $contato['id']; ?>">[ EXCLUIR ]</a>
-                </td>
+                <td colspan="4" align="center">Nenhum registro encontrado!</td>
             </tr>
-        <?php endforeach; ?>
+        <?php endif; ?>
     </tbody>
 </table>
